+++ eco-application/eco-application-multi_tenant/src/test/kotlin/community/flock/eco/application/multitenant/RegisterTenantTest.kt
@@ -1,7 +1,8 @@
-package community.flock.eco.application.multitenant
+package community.flock.eco.application.multitenant.community.flock.eco.application.multitenant
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import community.flock.eco.application.multitenant.controllers.RegistrationInput
+import community.flock.eco.feature.multitenant.services.MultiTenantSchemaService
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
@@ -11,44 +12,24 @@ import org.springframework.mock.web.MockHttpSession
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
 import org.springframework.test.web.servlet.result.MockMvcResultHandlers
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers.content
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 import java.util.UUID
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
-class CrossLoginTenantTest {
+class RegisterTenantTest {
     @Autowired
     lateinit var mockMvc: MockMvc
+
+    @Autowired
+    lateinit var multiTenantSchemaService: MultiTenantSchemaService
 
     @Autowired
     lateinit var objectMapper: ObjectMapper
 
     @Test
-    fun `should not be able to cross login into tenant`() {
-        val sessionA = MockHttpSession()
-        val sessionB = MockHttpSession()
-
-        val tenantA = register(sessionA)
-        val tenantB = register(sessionB)
-
-        login(sessionA, tenantA)
-        login(sessionB, tenantB)
-
-        me(sessionA, tenantA)
-        me(sessionB, tenantB)
-
-        mockMvc.perform(
-            MockMvcRequestBuilders
-                .get("/api/users/me")
-                .session(sessionA)
-                .header("X-TENANT", tenantB.tenantName),
-        )
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().isUnauthorized)
-    }
-
-    fun register(session: MockHttpSession): RegistrationInput {
+    fun registerTenant() {
         val random = UUID.randomUUID().toString().replace("-", "_")
         val input =
             RegistrationInput(
@@ -56,6 +37,8 @@ class CrossLoginTenantTest {
                 email = "tenant@$random.nl",
                 name = "Tenant de Tenant",
             )
+
+        val session = MockHttpSession()
 
         mockMvc.perform(
             MockMvcRequestBuilders
@@ -69,36 +52,29 @@ class CrossLoginTenantTest {
             .andDo(MockMvcResultHandlers.print())
             .andExpect(status().isOk)
 
-        return input
-    }
-
-    fun login(
-        session: MockHttpSession,
-        tenant: RegistrationInput,
-    ) {
         mockMvc.perform(
             MockMvcRequestBuilders
                 .post("/login")
                 .session(session)
-                .header("X-TENANT", tenant.tenantName)
-                .content("username=${tenant.email}&password=password")
+                .header("X-TENANT", input.tenantName)
+                .content("username=${input.email}&password=password")
                 .contentType(MediaType.APPLICATION_FORM_URLENCODED),
         )
             .andDo(MockMvcResultHandlers.print())
             .andExpect(status().is3xxRedirection)
-    }
 
-    fun me(
-        session: MockHttpSession,
-        tenant: RegistrationInput,
-    ) {
         mockMvc.perform(
             MockMvcRequestBuilders
                 .get("/api/users/me")
                 .session(session)
-                .header("X-TENANT", tenant.tenantName),
+                .header("X-TENANT", input.tenantName),
         )
             .andDo(MockMvcResultHandlers.print())
             .andExpect(status().isOk)
+            .andExpect(
+                MockMvcResultMatchers
+                    .jsonPath("$.email")
+                    .value(input.email),
+            )
     }
 }
