+++ eco-cloud/eco-cloud-stub/src/test/kotlin/community/flock/eco/cloud/stub/StubStorageServiceTest.kt
